Wire up validation pipeline: CI + GIAB/fuzz/conformance harnesses (manuscript P0/P1)

+++ pkg/htsgo/baq/capmapq_test.go
@@ -18,7 +18,7 @@ func makeRec(pos int32, cigar, seq string, q byte) *sam.Record {
 		qual[i] = q
 	}
 	return &sam.Record{
-		QName: "r", RName: "c", Pos: pos, MapQ: 60,
+		QName: "r", RName: "c", Pos: int64(pos), MapQ: 60,
 		Cigar: c, Seq: seq, Qual: qual,
 	}
 }
